fix: clippy errors

+++ onnx-tracer/src/tensor/ops.rs
@@ -1705,10 +1705,8 @@ pub fn argmax_axes<T: TensorType + Add<Output = T> + std::cmp::Ord + From<i32> +
 /// ).unwrap();
 /// assert_eq!(result, expected);
 /// ```
-
-// TODO(AntoineF4C5): generic quantization
-// Ensure T implements `From<u64>`
 pub fn argmin_axes<T: TensorType + Add<Output = T> + std::cmp::Ord + From<i32> + Send + Sync>(
+    // TODO(AntoineF4C5): generic quantization
     a: &Tensor<T>,
     dim: usize,
 ) -> Result<Tensor<T>, TensorError> {
@@ -2009,9 +2007,8 @@ pub fn intercalate_values<T: TensorType>(
 ///                                           0, 0, 0, 0, 1]), &[2, 2, 5]).unwrap();
 /// assert_eq!(result, expected);
 /// ```
-
-// TODO(AntoineF4C5): generic quantization
 pub fn one_hot(
+    // TODO(AntoineF4C5): generic quantization
     tensor: &Tensor<i32>,
     num_classes: usize,
     axis: usize,
@@ -2308,9 +2305,8 @@ pub fn deconv<
 /// let expected: Tensor<i32> = Tensor::<i32>::new(Some(&[3, 2, 2, 3]), &[1, 1, 2, 2]).unwrap();
 /// assert_eq!(pooled, expected);
 /// ```
-
-// TODO(AntoineF4C5): generic quantization
 pub fn sumpool(
+    // TODO(AntoineF4C5): generic quantization
     image: &Tensor<i32>,
     padding: [(usize, usize); 2],
     stride: (usize, usize),
@@ -2738,9 +2734,8 @@ pub fn slice<T: TensorType + Send + Sync>(
 // ---------------------------------------------------------------------------------------------------------
 
 /// Activation functions
-
-// TODO(AntoineF4C5): generic quantization
 pub mod nonlinearities {
+    // TODO(AntoineF4C5): generic quantization
     use super::*;
 
     /// Ceiling operator.
@@ -3747,16 +3742,15 @@ pub mod nonlinearities {
     /// let expected = Tensor::<i32>::new(Some(&[1, 0, 1, 3, 0, 0]), &[2, 3]).unwrap();
     /// assert_eq!(result, expected);
     /// ```
-
-    // TODO(AntoineF4C5): generic quantization
     pub fn const_div(a: &Tensor<i32>, denom: f64) -> Tensor<i32> {
+        // TODO(AntoineF4C5): generic quantization
         a.par_enum_map(|_, a_i| {
-            let mut d_inv_x = (a_i as i32) / (denom as i32);
-            let remainder = a_i as i32 % denom as i32;
+            let mut d_inv_x = a_i / (denom as i32);
+            let remainder = a_i % denom as i32;
             if (remainder < 0 && (denom as i32) > 0) || (remainder > 0 && (denom as i32) < 0) {
                 d_inv_x -= 1;
             }
-            Ok::<_, TensorError>(d_inv_x as i32)
+            Ok::<_, TensorError>(d_inv_x)
         })
         .unwrap()
     }
@@ -3915,9 +3909,8 @@ pub mod nonlinearities {
     /// let expected = Tensor::<i32>::new(Some(&[2]), &[1]).unwrap();
     /// assert_eq!(result, expected);
     /// ```
-
-    // TODO(AntoineF4C5): generic quantization
     pub fn mean(a: &Tensor<i32>, scale: usize) -> Tensor<i32> {
+        // TODO(AntoineF4C5): generic quantization
         let sum = sum(a).unwrap();
         const_div(&sum, (scale * a.len()) as f64)
     }
@@ -3941,9 +3934,8 @@ pub mod nonlinearities {
     /// ).unwrap();
     /// assert_eq!(result, expected);
     /// ```
-
-    // TODO(AntoineF4C5): generic quantization
     pub fn mean_of_squares_axes(a: &Tensor<i32>, axes: &[usize]) -> Tensor<i32> {
+        // TODO(AntoineF4C5): generic quantization
         let square = a.map(|a_i| a_i * a_i);
         let sum = sum_axes(&square, axes).unwrap();
         let denominator = a.len() / sum.len();
